+++ vo08/login/public/index.php
@@ -2,9 +2,11 @@
 
 declare(strict_types=1);
 
+use Fhooe\Latte\RouterExtension;
+use Fhooe\Latte\SessionExtension;
 use Fhooe\Router\Router;
-use Fhooe\Twig\RouterExtension;
-use Fhooe\Twig\SessionExtension;
+use Latte\Engine;
+use Latte\Loaders\FileLoader;
 use LoginExample\CreateDB;
 use LoginExample\Login;
 use LoginExample\PasswordHash;
@@ -13,9 +15,6 @@ use Monolog\Formatter\LineFormatter;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Monolog\Processor\PsrLogMessageProcessor;
-use Twig\Environment;
-use Twig\Extension\DebugExtension;
-use Twig\Loader\FilesystemLoader;
 
 require "../vendor/autoload.php";
 
@@ -32,7 +31,7 @@ session_start([
 ]);
 
 /**
- * Instantiated Router invocation. Create an object, define the routes and run it.
+ * Instantiated Router invocation. Create an object, define the routes, and run it.
  */
 // Create a monolog instance for logging in the skeleton.
 $logger = new Logger("skeleton-logger");
@@ -50,75 +49,68 @@ $logger->pushHandler($handler);
 // Create a new Router object with the logger.
 $router = new Router($logger);
 
-// Create a new Twig instance for advanced templates.
-$twig = new Environment(
-    new FilesystemLoader("../views"),
-    [
-        "cache" => "../cache",
-        "auto_reload" => true,
-        "debug" => true,
-    ],
-);
-
-// Add the router extension to Twig. This makes the url_for() and get_base_path() functions available in templates.
-$twig->addExtension(new RouterExtension($router));
-// Add the session extension to Twig. This makes the session() function available in templates to access entries in $_SESSION.
-$twig->addExtension(new SessionExtension());
-// Add the debug extension to Twig. This makes the dump() function available in templates to dump variables.
-$twig->addExtension(new DebugExtension());
+// Create a new Latte instance for advanced templates.
+$latte = new Engine();
+$latte->setLoader(new FileLoader(__DIR__ . "/../views"));
+$latte->setCacheDirectory(__DIR__ . "/../cache");
+$latte->addExtension(new RouterExtension($router));
+$latte->addExtension(new SessionExtension());
 
 // Set a base path if your code is not in your server's document root.
-$router->setBasePath("/hyp2vo-t1-examples/vl07/login/public");
+$router->basePath = "/wet2vo-examples/vo08/login/public";
 
-// Set a 404 callback that is executed when no route matches.
-// Example for the use of an arrow function. It automatically includes variables from the parent scope (such as $twig).
-$router->set404Callback(fn() => $twig->display("404.html.twig"));
+// Set a 404 callback executed when no route matches.
+// Example for the use of an arrow function. It automatically includes variables from the parent scope (such as $latte).
+$router->set404Callback(fn() => $latte->render("404.latte"));
 
 // Define all routes here.
-$router->get("/", function () use ($twig) {
-    $twig->display("index.html.twig");
+$router->get("/", function () use ($latte) {
+    $latte->render("index.latte");
 });
 
-$router->get("/login", function () use ($twig, $router) {
+$router->get("/login", function () use ($latte, $router) {
     RouteGuard::requireNotLoggedIn("loginToken", $router, "/main");
-    $twig->display("login.html.twig");
+    $latte->render("login.latte");
 });
 
-$router->post("/login", function () use ($twig, $router) {
-    $login = new Login($twig, $router);
+$router->post("/login", function () use ($latte, $router) {
+    $login = new Login($latte, $router);
     $login->isValid();
     $login->displayOutput();
 });
 
-$router->get("/main", function () use ($twig, $router) {
+$router->get("/main", function () use ($latte, $router) {
     RouteGuard::requireLoggedIn("loginToken", $router, "/login");
-    $twig->display("main.html.twig");
+    $latte->render("main.latte");
 });
 
 $router->get("/logout", function () use ($router) {
     $_SESSION = [];
     if (isset($_COOKIE[session_name()])) {
-        setcookie(session_name(), "", time() - 86400, "/");
+        setcookie(session_name(), "", [
+            "expires" => time() - 86400,
+            "path"    => "/",
+        ]);
     }
     session_destroy();
     $router->redirectTo("/");
 });
 
-$router->get("/createdb", function () use ($twig) {
-    $createDB = new CreateDB($twig);
+$router->get("/createdb", function () use ($latte) {
+    $createDB = new CreateDB($latte);
     $createDB->createSchema();
     $createDB->createTable();
     $createDB->addUsers();
     $createDB->displayOutput();
 });
 
-$router->get("/passwordhash", function () use ($twig) {
-    $generatePassword = new PasswordHash($twig);
+$router->get("/passwordhash", function () use ($latte) {
+    $generatePassword = new PasswordHash($latte);
     $generatePassword->displayOutput();
 });
 
-$router->post("/passwordhash", function () use ($twig) {
-    $generatePassword = new PasswordHash($twig);
+$router->post("/passwordhash", function () use ($latte) {
+    $generatePassword = new PasswordHash($latte);
     $generatePassword->displayOutput();
 });