+++ vo08/login/src/LoginExample/RouteGuard.php
@@ -8,8 +8,8 @@ use Fhooe\Router\Router;
  * RouteGuard is a simple, session-based mechanism to protect certain routes. It can either check if a session token is
  * present and if not, perform a redirect or test if a session token is explicitly not present.
  * @package LoginExample
- * @author Wolfgang Hochleitner <wolfgang.hochleitner@fh-hagenberg.at>
- * @version 2025
+ * @author  Wolfgang Hochleitner <wolfgang.hochleitner@fh-hagenberg.at>
+ * @version 2026
  */
 class RouteGuard
 {
@@ -30,7 +30,6 @@ class RouteGuard
 
     /**
      * Requires that a session token is not present to access the route. If this is not the case,
-     * a redirect is performed to the route provided in $exitRoute.
      * @param string $tokenName The name of the session token to check for.
      * @param Router $router The Router object to perform the redirect.
      * @param string $exitRoute The route to redirect if the session token is present valid.